Run in full screen on startup, and clean init file
On OSX I run most things in full screen, so it makes sense to open in
fullscreen by default. The only time I can see this being a problem is
when multiple frames are taken into account, but we will see.

Also I have 'cleaned' the .spacemacs file, removing most comments, the
variable names are descriptive enough, ans should they not be help is
easy to find.